RenderTarget + BlendState + continued refactoring
- Created `RenderTarget` class
- Created `BlendState` class
- Continued refactoring
---- Separated header and source files into folders
----- Using _lowerCamelCase for private and protected member fields and functions
----- Move ctors for some types, more work is required for other classes using ComPtr

+++ 3DRadSpace/3DRadSpace_Editor_WindowsDX11/EditObjectDialog.cpp
@@ -785,7 +785,7 @@ bool EditObjectDialog::setObject()
 				case FieldRepresentationType::Key:
 				{
 					HWND hotkey = std::get<HWND>(windows[i++]);
-					int r = SendMessageA(hotkey, HKM_GETHOTKEY, 0, 0);
+					int r = (int)SendMessageA(hotkey, HKM_GETHOTKEY, 0, 0);
 					Input::Key k = static_cast<Input::Key>(LOBYTE(LOWORD(r)));
 					if(k == Input::Key::None)
 					{
